requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyMySQL
bcrypt
PyJWT
python-dotenv
requests>=2.32.0
tenacity>=8.2.0
python-dateutil>=2.9.0